worker/requirements.txt
boto3==1.34.0
Pillow
requests==2.31.0
python-dateutil==2.9.0.post0
